+++ lib/src/api/src/gql_builder/union.dart
@@ -1,0 +1,13 @@
+part of '../../api.dart';
+
+class GqlUnion extends GqlField {
+  GqlUnion({
+    required super.name,
+    super.fields,
+  });
+
+  @override
+  String toString() {
+    return 'GqlUnion(name: $name, fields: $fields)';
+  }
+}